vc10 compatibility + xerces 3.1
git-svn-id: http://svn.cip4.org/svnroot/trunk@4751 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/jdf/io/FileInputStream.cpp
@@ -273,7 +273,7 @@ JDFUInt64 FileInputStream::length()
 
 JDFUInt64 FileInputStream::seek(JDFUInt64 pos, int moveMethod)
 {
-	PlatformUtils::seekFile(m_handle,pos,moveMethod);
+	PlatformUtils::seekFile(m_handle,(int)pos,moveMethod);
 	return PlatformUtils::curFilePos(m_handle);
 }
